+++ src/libromdata/RomDataFactory.cpp
@@ -32,6 +32,7 @@ using std::unordered_set;
 using std::vector;
 
 // RomData subclasses: Consoles
+#include "Console/Atari7800.hpp"
 #include "Console/CBMCart.hpp"
 #include "Console/Dreamcast.hpp"
 #include "Console/DreamcastSave.hpp"
@@ -243,6 +244,7 @@ pthread_once_t RomDataFactoryPrivate::once_mimeTypes = PTHREAD_ONCE_INIT;
 // TODO: Add support for multiple magic numbers per class.
 const RomDataFactoryPrivate::RomDataFns RomDataFactoryPrivate::romDataFns_magic[] = {
 	// Consoles
+	GetRomDataFns_addr(Atari7800, ATTR_HAS_METADATA, 4, 'RI78'),	// "ATARI7800"
 	GetRomDataFns_addr(PlayStationEXE, 0, 0, 'PS-X'),
 	GetRomDataFns_addr(SufamiTurbo, ATTR_HAS_THUMBNAIL | ATTR_HAS_METADATA, 8, 'FC-A'),	// Less common than "BAND"
 	GetRomDataFns_addr(WiiU, ATTR_HAS_THUMBNAIL | ATTR_SUPPORTS_DEVICES, 0, 'WUP-'),